Items are now two 16 Bit indices.

+++ src/saiga/opengl/rendering/lighting/cpu_plane_clusterer.h
@@ -41,7 +41,7 @@ class SAIGA_OPENGL_API CPUPlaneClusterer : public Clusterer
     std::vector<Plane> planesZ;
 
     int avgAllowedItemsPerCluster = 128;
-    std::vector<std::vector<int>> clusterCache;
+    std::vector<std::vector<int16_t>> clusterCache;
 
     bool refinement = true;
